--- src/main/java/com/github/ethanicuss/astraladditions/blocks/DesizerBlockTemp.java
@@ -1,43 +1,0 @@
-package com.github.ethanicuss.astraladditions.blocks;
-
-import com.github.ethanicuss.astraladditions.registry.ModBlocks;
-import net.fabricmc.fabric.api.event.lifecycle.v1.ServerChunkEvents;
-import net.minecraft.block.AbstractBlock;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.server.world.ServerWorld;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
-import net.minecraft.world.chunk.Chunk;
-
-import java.util.HashSet;
-import java.util.Set;
-
-//This is to be removed in a later update. Along with anything else using it.
-//This is here to allow the blocks to change to the new casing
-
-public class DesizerBlockTemp extends Block  {
-    public DesizerBlockTemp(Settings settings) {
-        super(settings);
-    }
-
-    @Override
-    public void onPlaced(World world, BlockPos pos, BlockState state, LivingEntity placer, ItemStack itemStack) {
-        super.onPlaced(world, pos, state, placer, itemStack);
-        if (!world.isClient) {
-            BlockState newBlockState = ModBlocks.DESIZER_BASE.getDefaultState();
-            world.setBlockState(pos, newBlockState);
-        }
-    }
-
-    @Override
-    public void onBlockAdded(BlockState state, World world, BlockPos pos, BlockState oldState, boolean notify) {
-        if (!world.isClient) {
-            BlockState newBlockState = ModBlocks.DESIZER_BASE.getDefaultState();
-            world.setBlockState(pos, newBlockState);
-        }
-    }
-
-}
