+++ src/main/java/com/github/ethanicuss/astraladditions/blocks/DesizerCasingBlock.java
@@ -1,13 +1,18 @@
 package com.github.ethanicuss.astraladditions.blocks;
 
+import com.github.ethanicuss.astraladditions.registry.ModBlocks;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.HorizontalFacingBlock;
+import net.minecraft.entity.LivingEntity;
 import net.minecraft.item.ItemPlacementContext;
+import net.minecraft.item.ItemStack;
 import net.minecraft.state.StateManager;
 import net.minecraft.state.property.EnumProperty;
 import net.minecraft.util.StringIdentifiable;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
+import net.minecraft.world.World;
 
 public class DesizerCasingBlock extends HorizontalFacingBlock {
 
@@ -17,6 +22,15 @@ public class DesizerCasingBlock extends HorizontalFacingBlock {
         super(settings);
         this.setDefaultState(this.stateManager.getDefaultState().with(FACING, Direction.NORTH).with(TYPE, Type.BASE));
 
+    }
+    //! This can be removed in a later update
+    @Override
+    public void onPlaced(World world, BlockPos pos, BlockState state, LivingEntity placer, ItemStack itemStack) {
+        super.onPlaced(world, pos, state, placer, itemStack);
+        if (!world.isClient) {
+            BlockState newBlockState = ModBlocks.DESIZER_BASE.getDefaultState();
+            world.setBlockState(pos, newBlockState);
+        }
     }
 
     @Override
