+++ src/main/resources/data/astraladditions/recipes/transmute.json
@@ -10,8 +10,8 @@
       "count": 1
     },
     {
-      "item": "minecraft:gold_ingot",
-      "count": 1
+      "item": "minecraft:iron_ingot",
+      "count": 2
     },
     {
       "item": "minecraft:dirt",
